move all clauses classes to Clauses dir
