+++ src/Grammar/Clauses/OrderByClause.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Abdulelahragih\QueryBuilder\Grammar;
+namespace Abdulelahragih\QueryBuilder\Grammar\Clauses;
 
 use Abdulelahragih\QueryBuilder\Types\OrderType;
